validator/control_node/requirements.txt
httpx~=0.27
python-dotenv~=1.0
redis~=5.0
colorama==0.4.6
asyncpg==0.29.0
asyncclick==8.1.7.2
rich==13.7.1
# numpy==1.26.4